requirements.txt
# 의존성 목록
PyYAML
scapy
